[MOVED] Minor refactor - Moved files and renamed packages
