+++ core-lib/src/main/java/info/hossainkhan/android/core/headlines/HeadlinesPresenter.java
@@ -38,8 +38,8 @@ import info.hossainkhan.android.core.base.BasePresenter;
 import info.hossainkhan.android.core.data.CategoryNameResolver;
 import info.hossainkhan.android.core.model.CardItem;
 import info.hossainkhan.android.core.model.NavigationRow;
-import info.hossainkhan.android.core.newssource.NewsProvider;
-import info.hossainkhan.android.core.newssource.NyTimesNewsProvider;
+import info.hossainkhan.android.core.model.NewsProvider;
+import info.hossainkhan.android.core.newsprovider.NyTimesNewsProvider;
 import io.swagger.client.ApiClient;
 import io.swagger.client.api.ConsumptionFormat;
 import io.swagger.client.api.StoriesApi;
