+++ core-lib/src/main/java/info/hossainkhan/android/core/model/NewsProvider.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package info.hossainkhan.android.core.newssource;
+package info.hossainkhan.android.core.model;
 
 import java.util.Set;
