+++ mobile/src/main/java/info/hossainkhan/dailynewsheadlines/MainActivity.java
@@ -15,8 +15,8 @@ import info.hossainkhan.android.core.headlines.HeadlinesContract;
 import info.hossainkhan.android.core.headlines.HeadlinesPresenter;
 import info.hossainkhan.android.core.model.CardItem;
 import info.hossainkhan.android.core.model.NavigationRow;
-import info.hossainkhan.android.core.newssource.NewsProvider;
-import info.hossainkhan.android.core.newssource.NyTimesNewsProvider;
+import info.hossainkhan.android.core.model.NewsProvider;
+import info.hossainkhan.android.core.newsprovider.NyTimesNewsProvider;
 import timber.log.Timber;
 
 public class MainActivity extends BaseActivity implements HeadlinesContract.View {
