+++ tv/src/main/java/info/hossainkhan/dailynewsheadlines/browser/HeadlinesBrowseFragment.java
@@ -39,8 +39,8 @@ import info.hossainkhan.android.core.headlines.HeadlinesContract;
 import info.hossainkhan.android.core.headlines.HeadlinesPresenter;
 import info.hossainkhan.android.core.model.CardItem;
 import info.hossainkhan.android.core.model.NavigationRow;
-import info.hossainkhan.android.core.newssource.NewsProvider;
-import info.hossainkhan.android.core.newssource.NyTimesNewsProvider;
+import info.hossainkhan.android.core.model.NewsProvider;
+import info.hossainkhan.android.core.newsprovider.NyTimesNewsProvider;
 import info.hossainkhan.android.core.util.UiUtils;
 import info.hossainkhan.dailynewsheadlines.R;
 import info.hossainkhan.dailynewsheadlines.browser.listeners.CardItemViewInteractionListener;
